move auth methods back to resolvable to reduce the diff
